requirements-test.txt
-r ./requirements.txt

Whoosh==2.5.5
coverage
coveralls
mock>=1.0.1
nose>=1.3.0
django-nose>=1.2
